backend/requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
django-debug-toolbar
python-decouple
django-redis
djangorestframework-simplejwt
pymupdf4llm
google-generativeai
openai
boto3
celery
flower
django-celery-results
Pillow
weasyprint
beautifulsoup4
langchain
langchain-community
pdfplumber
pypdf
tiktoken
chromadb
fastapi
uvicorn
python-dotenv
redis
pydantic 